+++ src/tests/integration/tabs.rs
@@ -1,9 +1,10 @@
 use insta::assert_snapshot;
 
 use crate::tests::fakes::FakeInputOutput;
+use crate::tests::start;
 use crate::tests::utils::{get_next_to_last_snapshot, get_output_frame_snapshots};
+use crate::CliArgs;
 use crate::{panes::PositionAndSize, tests::utils::commands::CLOSE_PANE_IN_PANE_MODE};
-use crate::{start, CliArgs};
 
 use crate::common::input::{config::Config, options::Options};
 use crate::tests::utils::commands::{
